Do not test stories, only test components
